+++ packages/cli/plugin-garfish/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/plugin-garfish
 
+## 1.3.1
+
+### Patch Changes
+
+- 53aca274: modify garfish-plugin config type
+- Updated dependencies [53aca274]
+- Updated dependencies [78279953]
+- Updated dependencies [e116ace5]
+- Updated dependencies [4d72edea]
+  - @modern-js/core@1.4.1
+  - @modern-js/utils@1.3.1
+
 ## 1.3.0
 
 ### Minor Changes